requirements.txt
# 核心库通过源码安装，这里只列出额外的
# ultralytics -> 源码已安装 (train_project/ultralytics_repo)
# opencv-python -> 源码安装时已包含
# torch -> 源码安装时已包含
# matplotlib -> 源码安装时已包含

numpy>=1.23.0
matplotlib>=3.3.0
opencv-python>=4.6.0
pillow>=7.1.2
pyyaml>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.8.0
torchvision>=0.9.0
psutil>=5.8.0
polars>=0.20.0
ultralytics-thop>=2.0.18
onnx>=1.12.0
onnxslim>=0.1.82
PyQt5  # 用于界面开发
